Improved accuracy of Count, sort and print (faded example)
This commit fixes two issues with the example.

First, the sort should be numerical, rather than lexical, both for consistency with previous examples and to solve the challenge.

Second, the `wc` command outputs a total row, which is numerically larger than all the others and therefore will appear as the last line. If you only look at the last 10 lines, you will see 9 files and the total, not the 10 files asked for in the challenge.