Update 05-counting-mining.md
In the faded example the -w flag is already introduced, but then we tell learners to use the man or --help commands a couple challenges later to find out about it. This doesn't really make sense considering we told them about it already, so I changed the faded example to include the -w flag and not the wc with the instructions to look up how to count words for that challenge. The later challenge I then rewrote just telling them to explore the -w flag more. You might want to get rid of one of the challenges or change them instead of fixing them how I have done it here.